Refactor: simplify OS's execution scheme
Removed unnecessary ExecutionInfo and put everything in ExecutionResult instead.

ExecutionResult now only has a string representing stderr instead of two istream*s.
Caller should open the output/error files explicitly to read them, and then close them.
This is to make it clear that the caller is responsible for closing the open files.